timer: move get_frame_info to utils
